% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf.test.R
\name{adf.test}
\alias{adf.test}
\title{Calculates ADF Stationarity Test}
\usage{
adf.test(y, type = c("none", "drift", "trend"), lags = 1,
  selectlags = c("Fixed", "AIC", "BIC"))
}
\arguments{
\item{y}{Data series}

\item{type}{The type of test. Whether it includes a drift or trend component.}

\item{lags}{Number of lags to test}

\item{selectlags}{Whether a model criterion should be used to determine a lag length.}
}
\description{
Calculates ADF Test Info based from the urca package
}
\author{
Nicholas Dregne and Helena Ristov
}

