% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aroon.R
\name{Aroon}
\alias{Aroon}
\title{Calculates the Aroon oscillator and outputs a signal for buy and sell events}
\usage{
Aroon(spread.data, vol.settings, volBand, aroon.settings,
  use.volFilter = TRUE)
}
\arguments{
\item{spread.data}{A price series}

\item{vol.settings}{A list with the vol settings to be used for the threshold computation}

\item{volBand}{The number of standard deviations that set the zone around which we compute the aroon values.}

\item{aroon.settings}{The time bucket for the Aroon oscillator and the threshold value for the buy and sell signal}

\item{use.volFilter}{filter out any values that are inside the volBand}
}
\value{
list of Aroon indicators and triggers:
\item{aroon.up.sig}{The value of the oscillator if the threshold on the upside is breached and outside the volBand, NA otherwise}
\item{aroon.down.sig}{The value of the oscillator if the threshold on the downside is breached and outside the volBand, NA otherwise}
\item{binary.signal}{The combinations of the up and down signal into one field with 0 implying no action, -1 implying a sell, and 1 implying a buy}
\item{binary.filter}{The binary signal with consecutive buy and sell signals filtered out}
}
\description{
Calculates the Aroon oscillator and outputs a signal for buy and sell events
}
\author{
Helena Ristov and Cameron Sawyer
}

