% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.OHLC.R
\name{Calc.OHLC}
\alias{Calc.OHLC}
\title{Creates an Open, High, Low, Close time-series for a given contract.}
\usage{
Calc.OHLC(data, contract, barLength)
}
\arguments{
\item{data}{MDR or TRS data.}

\item{contract}{The contract that you are pulling data for.}

\item{barLength}{The length of bar for OHLC calculation in seconds.}
}
\description{
Creates an Open, High, Low, Close time-series for a given contract.
}
\author{
Helen Ristov and Cameron Sawyer
}

