% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.Sinusoidal.FV.R
\name{Calc.Sinusoidal.FV}
\alias{Calc.Sinusoidal.FV}
\title{Calculates a low amplitude sinusoidal regression based on a spectral analysis}
\usage{
Calc.Sinusoidal.FV(data, window, incr, anchor.window = FALSE)
}
\arguments{
\item{data}{A price series}

\item{window}{The window of time for the spectral fit which can be rolling or anchored}

\item{incr}{The increment of time that the prices series is squared at}

\item{anchor.window}{The anchor window that will be used to start the cumulative fit}
}
\value{
list of values are returned
\item{lastValue}{The last fitted FV}
\item{fitted.values}{The entire range of fitted fair values }
}
\description{
Calculates a low amplitude sinusoidal regression based on a spectral analysis
}
\author{
Helena Ristov and Nicholas Dregne
}

