% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.SMA.R
\name{Calc.SMA}
\alias{Calc.SMA}
\title{Creates start and end datetimes for pullling data}
\usage{
Calc.SMA(data, contract, start.date, end.date, incr, window.size,
  window.size.units)
}
\arguments{
\item{data}{MDR or TRS data.}

\item{contract}{The contract that you are pulling data for.}

\item{start.date}{The start date of the SMA}

\item{end.date}{The end date of the SMA}

\item{incr}{Time bucket increment of SMA obs}

\item{window.size}{Number of units in SMA}

\item{window.size.units}{The unit of measure for the SMA length}
}
\description{
Creates start and end datetimes for pullling data
}
\author{
Helen Ristov
}
