% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.Sweep.Vol.Per.R
\name{Calc.Sweep.Vol.Per}
\alias{Calc.Sweep.Vol.Per}
\title{Calculate the Inflation Factor From Market Sweeps.}
\usage{
Calc.Sweep.Vol.Per(contract, start, end, book.per, Limit, VML)
}
\arguments{
\item{contract}{Contract to get sweep inflation factor.}

\item{start}{Start Date for analytical period.}

\item{end}{End Date for analytical period}

\item{book.per}{Percentile of book size that is considered liquid.}
}
\description{
Calculate the Inflation Factor From Market Sweeps.
}
\author{
Helen Ristov and Nic Dregne
}

