% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc.VWAP.R
\name{Calc.VWAP}
\alias{Calc.VWAP}
\title{Creates a VWAP series for a given contract}
\usage{
Calc.VWAP(data, contract, lookBackPeriod)
}
\arguments{
\item{data}{MDR or TRS data.}

\item{contract}{The contract that you are pulling data for.}

\item{lookBackPeriod}{The lookback period of rolling VWAP calculation in seconds.}
}
\description{
Creates a VWAP series for a given contract
}
\author{
Helen Ristov and Cameron Sawyer
}

