% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBidAsk.Edge.R
\name{calcBidAsk.Edge}
\alias{calcBidAsk.Edge}
\title{Calculates Perceived Market Edge}
\usage{
calcBidAsk.Edge(market.data, contract, start.lean = NULL, end.lean = NULL,
  edge.floor = NULL, min.edge.qty = NULL, max.edge.qty = NULL,
  TT_MinTick = FALSE)
}
\arguments{
\item{market.data}{calcMarketMeasures data set}

\item{contract}{Contract for which you desire to get edge measures.}

\item{start.lean}{Start lean value to determine lean factor for edge calc.}

\item{end.lean}{End lean value to determine lean factor for edge calc.}

\item{edge.floor}{Edge floor to use if we desire one.}

\item{min.edge.qty}{Min qty for edge floor calc.}

\item{max.edge.qty}{Max qty for edge floor calc.}

\item{TT_MinTick}{A boolean to determine if we should use the TT mintick or TRS mintick.}
}
\description{
Calculates Perceived Market Edge
}
\author{
Nicholas Dregne
}
