% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCovarInfo.R
\name{CalcCovarInfo}
\alias{CalcCovarInfo}
\title{Calculates Covariates for EQPos and Risk Redux Calcs}
\usage{
CalcCovarInfo(cov, spread, raw = FALSE, date = last(GetWeekDays(Sys.Date() -
  8, Sys.Date() - 1)))
}
\arguments{
\item{cov}{Covariance matrix from which covariates will be derived.}

\item{spread}{Spread for which to get covariates.}

\item{raw}{Optional parameter to return raw covarinputs for contracts with an non-one front beta.}
}
\description{
Calculates Covariates for EQPos and Risk Redux Calcs
}
\author{
Nicholas Dregne
}

