% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDrift.R
\name{calcDrift}
\alias{calcDrift}
\title{Calculates a drift measure for a price series}
\usage{
calcDrift(spread, dPeriod, .combine = TRUE)
}
\arguments{
\item{spread}{A price series on which you want to compute the drift}

\item{dPeriod}{The number of days used to compute the drift}

\item{.combine}{A boolean that sets if we want to merge the periods and return the mean of the rolling drift measures over those periods.}
}
\description{
Calculates a drift measure for a price series
}
\author{
Helena Ristov
}
