% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEff.BidAsk.R
\name{calcEff.BidAsk}
\alias{calcEff.BidAsk}
\title{Calculates Effective Bid Ask Prices}
\usage{
calcEff.BidAsk(market.data, createQty, cancelQty)
}
\arguments{
\item{market.data}{An MDR market data set.}

\item{createQty}{Start lean value to determine lean factor for edge calc.}

\item{cancelQty}{End lean value to determine lean factor for edge calc.}
}
\description{
Calculates Effective Bid Ask Prices
}
\author{
Nicholas Dregne
}

