% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEQPos.R
\name{CalcEQPos}
\alias{CalcEQPos}
\title{Calculates Equivalent Positions Based on an Outright Covariance Matrix}
\usage{
CalcEQPos(cov, pos, types)
}
\arguments{
\item{cov}{Outright covariance matrix to be used in the variance calculations}

\item{pos}{A vector of outright positions for which we wish to know the equivalent positions.}

\item{types}{Types of combinations to be calculated from the outright covariance matrix.}
}
\description{
Calculates EQ Pos for Pre-Selected Contract Types
}
\author{
Nicholas Dregne and Helena Ristov
}
