% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMarket.Measures.R
\name{calcMarket.Measures}
\alias{calcMarket.Measures}
\title{Calculates Standard Market Micro-Structures (WMP, MP, Bid-Ask)}
\usage{
calcMarket.Measures(market.data, contract, TT_MinTick = FALSE)
}
\arguments{
\item{market.data}{Raw TOB market data from either TRS or MDR}

\item{contract}{Contract for which you desire to get market micro-structure measures.}

\item{TT_MinTick}{A boolean to determine if we should use the TT mintick or TRS mintick.}
}
\description{
Calculates Standard Market Micro-Structures (WMP, MP, Bid-Ask)
}
\author{
Nicholas Dregne
}
