% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPC.Vector.R
\name{calcPC.Vector}
\alias{calcPC.Vector}
\title{Create PC Component Variable}
\usage{
calcPC.Vector(returns, dim = NULL, evMatrix = NULL)
}
\arguments{
\item{returns}{A stationary price series whose variation through time we wish to examine.}

\item{dim}{Optional input for the eigenvectors for which we wish to calculate the PC variable.  If left null, then all possible vector variables will be return.}

\item{evMatrix}{An optional eigenvector matrix to use in generating the PC variables.}
}
\description{
Calculates and records variables denoting the desired PC Eigenvectors' fluctuations.
}
\details{
Using an given price series and an option eigenvector matrix (will create it from price series otherwise), this function 
will calculate the principal components' variation throughout the price series.
}
\author{
Nicholas Dregne
}

