% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPSI.R
\name{calcPSI}
\alias{calcPSI}
\title{Calculates the population stability index for two sample variables}
\usage{
calcPSI(data, var, valvar, breaks = 10)
}
\arguments{
\item{data}{The dataset that contains the two variables for comparison}

\item{var}{The development variable or score used to set the distribution}

\item{valvar}{The validation variable to compare against the development variable}

\item{breaks}{The number of bins to break the development variable for stability testing. Default is decile distribution.}
}
\description{
Calculates the population stability index for two sample variables
}
\examples{
\dontrun{
x1      <- runif(1000, 1, 10)
x1_val  <- runif(1000, 1, 10)

random <- sample(c(0,1),1)
for(i in 1:999){
randomnext <- sample(c(0,1),1)
random     <- c(random, randomnext)
 
}

test <- cbind(x1, x1_val, random)
calPSI(test, 'x1', 'x1_val')
}
}
\author{
Helena Ristov
}
