% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRange.R
\name{calcRange}
\alias{calcRange}
\title{Calculates the range of a spread computed by percentile bands to minimize the impact of outliers}
\usage{
calcRange(spread, dPeriod = NA, minProb, maxProb, .combine = TRUE)
}
\arguments{
\item{spread}{A price series on which you want to compute the range of a spread}

\item{dPeriod}{The number of days for the period. The default is set to NA with .combine = true to compute the range over the entire spread}

\item{minProb}{The percentile to set the lower band of the range calculation}

\item{maxProb}{The percentile to set the upper band of the range calculation}

\item{.combine}{A boolean that sets if want to compute the range over the entire spread}
}
\description{
Calculates the range of a spread computed by percentile bands to minimize the impact of outliers
}
\author{
Helena Ristov
}
