% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcRiskRedux.R
\name{CalcRiskRedux}
\alias{CalcRiskRedux}
\title{Calculates Amount of Risk Reduced Based on an Outright Covariance Matrix}
\usage{
CalcRiskRedux(cov, pos, types, delta, ref.qty = 1, ref.contract = NULL)
}
\arguments{
\item{cov}{Outright covariance matrix to be used in the variance calculations}

\item{pos}{A vector of outright positions for which we wish to know the equivalent positions.}

\item{types}{Types of combinations to be calculated from the outright covariance matrix.}

\item{ref.qty}{Reference quantity used to determine what a full traded level would do to risk.}

\item{ref.contract}{The contract that should be used as the basis for reference volatility.}
}
\description{
Calculates Portfolio Risk Reduction for Pre-Selected Contract Types
}
\author{
Nicholas Dregne and Helena Ristov
}

