% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSemiVar.R
\name{calcSemiVar}
\alias{calcSemiVar}
\title{Calculates the SemiVariance of Price Series}
\usage{
calcSemiVar(data, split.num = NULL, n = NULL)
}
\arguments{
\item{data}{A price series on which you want to compute the semivariance}

\item{split.num}{The number to split the price series distribution.  If null, the mean is used.}

\item{n}{The number of observations to use in the calculation.  If null, the total set is combined.}
}
\description{
Calculates the SemiVariance of Price Series
}
\author{
Nicholas Dregne
}
