% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvergenceMetrics.R
\name{convergenceMetrics}
\alias{convergenceMetrics}
\title{Convergence test based on a disjoint distance from the limit value}
\usage{
convergenceMetrics(data, limit, epsilon, threshold, perf.period)
}
\arguments{
\item{data}{xts series to be tested for convergence (must be nx1)}

\item{limit}{value at which the series converges to}

\item{epsilon}{distance from limit which sets the convergence zone}

\item{threshold}{value above and below the limit which determine if an observation is disjoint}

\item{perf.period}{number of bars to test for convergence}
}
\value{
disjoint the distance the current observation is above or below the current threshold

convBool boolean to indicate if the series converged at some point during the performance period
}
\description{
Convergence test based on a disjoint distance from the limit value
}
\author{
Helena Ristov
}
