% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCov2ORCov.R
\name{CSCov2ORCov}
\alias{CSCov2ORCov}
\title{Transform Calendar Spread Covariance to Outright Covariance}
\usage{
CSCov2ORCov(CS_CovMat, ContractSymbol)
}
\arguments{
\item{CS_CovMat}{The calendar spread covariance matrix where the first contract is the single outright required.}

\item{ContractSymbol}{The symbol of the underlying asset class.}
}
\description{
Taking a covariance matrix of a single outright and multiple calendar spreads, this function will create a covariance matrix of all outrights.
}
\details{
This function, using the covariance identities, recreates what all outright covariances are based on the covariances
of the contracts' calendar spreads which tend to be more liquid and less volatile.  Note: Only one asset class can be used with this function.
}
\author{
Nicholas Dregne and Helena Ristov
}
