% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPullDates.R
\name{dataPullDates}
\alias{dataPullDates}
\title{Creates start and end datetimes for pullling data}
\usage{
dataPullDates(contract, tradeDays, endDate = Sys.Date() - 1, type = "A")
}
\arguments{
\item{contract}{The contract that you are pulling data for.}

\item{tradeDays}{The number of trade days of data to pull.}

\item{endDate}{The last day you wish to pull data for. Defaulted to yesterday.}
}
\description{
Creates start and end datetimes for pullling data
}
\author{
Cameron Sawyer
}
