% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePortfolio.R
\name{generatePortfolio}
\alias{generatePortfolio}
\title{Generates a portfolio for a given list of assets}
\usage{
generatePortfolio(Asset, Type, years, rev = FALSE)
}
\arguments{
\item{Asset}{a list of asset classes}

\item{Type}{A list of contract types that you want to generate for each asset class}

\item{years}{A list containing the number of years for each asset class}

\item{rev}{A reverse boolean that indicates the direction in which you want to pull the contracts. The default is set to FALSE for forward looking. Set to TRUE for building a portfolio in the past.}
}
\description{
Generates a portfolio for a given list of assets
}
\author{
Helena Ristov
}
