
% Generated by roxygen2 (4.0.1): do not edit by hand

% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IncrCounter.R

\name{IncrCounter}
\alias{IncrCounter}
\title{Incr Counter}
\usage{
IncrCounter(Prices, incr)
}
\arguments{
\item{Prices}{The time-series of prices.}

\item{incr}{The desired increment level.}
}
\description{
Calculates the number of times a time-series of prices has moved a pre-specified increment.
}
\details{
Given a time-series of prices and an increment level, the function will determine how many
times the price series has moved the specified increment level before reverting back a full
increment level.
}
\author{
Nicholas Dregne
}

