% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncrPerWindow.R
\name{IncrPerWindow}
\alias{IncrPerWindow}
\title{Calculates the number of increments for a specified window of time}
\usage{
IncrPerWindow(start.date, end.date, incr, window.size, window.size.units)
}
\arguments{
\item{start.date}{The beginning of the data pull window}

\item{end.date}{The ending of the data pull window}

\item{incr}{The bar length of the data pull}

\item{window.size}{The total length of the window in the units specified}

\item{window.units}{The time units. Supported values are secs, mins, hours, days}
}
\description{
Calculates the number of increments for a specified window of time
}
\author{
Helena Ristov
}
