
% Generated by roxygen2 (4.0.1): do not edit by hand

% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IncrScalpProfit.R

\name{IncrScalpProfit}
\alias{IncrScalpProfit}
\title{Volatilty Scalp Profit}
\usage{
IncrScalpProfit(TickCount, MaxLevels, slippage = 0)
}
\arguments{
\item{TickCount}{An increment count series.}

\item{MaxLevels}{Maximum number of levels that can be taken on}

\item{slippage}{Optional factor that penalizes the assumption of perfect execution by a percentage of the increment}
}
\description{
Internal function that calculates an increment profit given a specific increment count
}
\author{
Nicholas Dregne
}

