% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numClusters.R
\name{numClusters}
\alias{numClusters}
\title{Calculates the optimal number of clusters using a cluster R^2 equivalent metric, namely, the between sum of squares over the total sum of squares metric.}
\usage{
numClusters(Data, clusterIter = 10, threshold = 0.01)
}
\arguments{
\item{Data}{The data that you wish to examine containing the cluster dimensions}

\item{clusterIter}{The number of clusters centers for examination. A setting of 10 will run 2 through 10 cluster functions and return the best choice.}

\item{threshold}{The difference in the cluster R^2 must meet in order to include the next cluster iteration.}
}
\description{
Calculates the optimal number of clusters using a cluster R^2 equivalent metric, namely, the between sum of squares over the total sum of squares metric.
}
\author{
Helena Ristov
}
