% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR2CSCov.R
\name{OR2CSCov}
\alias{OR2CSCov}
\title{Outright to Calendar Spread Covariance Converter}
\usage{
OR2CSCov(ORCov)
}
\arguments{
\item{ORCov}{A covariance matrix of outright contracts.}
}
\description{
Convert Outright Covariance to a Calendare Spread Covariance
}
\details{
Takes an outright covariance with consecutive contracts and creates the calendar spread
covariance matrix
}
\author{
Nicholas Dregne and Helena Ristov
}

