% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBands.R
\name{pBands}
\alias{pBands}
\title{Calculates the pband levels on a spread price series}
\usage{
pBands(spread, dFast, dSlow, minFact, maxFact)
}
\arguments{
\item{spread}{A price series on which you want to compute the probability bands.}

\item{dFast}{The number of days used to compute the fast moving average}

\item{dSlow}{The number of days used to compute the slow moving average}

\item{minFact}{The number of standard deviations to set the lower bound of the scalping bounds}

\item{maxFact}{The number of standard deviations to set the upper bound of the scalping bounds}
}
\description{
Calculates the pband levels on a spread price series
}
\author{
Helena Ristov
}

