% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Create.R
\name{PCA_Create}
\alias{PCA_Create}
\title{Create PCA Matrix}
\usage{
PCA_Create(OR_Cov, Dimension)
}
\arguments{
\item{OR_Cov}{A covariance matrix of outright contracts.}

\item{Dimension}{The number of eigenvectors to return.}
}
\description{
Calculates and records a matrix comprising of the principal component eigenvalues and
eigenvectors.
}
\details{
Using an given outright covariance and a specified number of dimensions to use, this function 
will calculate the principal components returning a matrix with the first column corresponding
to the eigenvalues and the remaining columns the prespecified number of eigenvectors.
}
\author{
Nicholas Dregne
}
