
% Generated by roxygen2 (4.0.1): do not edit by hand

% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ScalpFVInd.R

\name{ScalpFVInd}
\alias{ScalpFVInd}
\title{Creates the indicator for a Spread Scalping Strategy}
\usage{
ScalpFVInd(Data, Vol, InitVolFactor, MaxVolFactor, AdjFactor)
}
\arguments{
\item{Data}{Single price series to generate a scalping FV indicator.}

\item{Vol}{The volatility measure for the given data above}

\item{InitVolFactor}{The multiplier of the Vol that determines the initial entry price.}

\item{MaxVolFactor}{The multiplier of the Vol that determines the max position price.}

\item{AdjFactor}{The amount that a price should be adjusted when the price exceeds the max price limit.}
}
\description{
Generates a fixed price scalp indicator for a given data set.  This inidcator is typically used to scalp around a stationary price series.
}
\author{
Nicholas Dregne
}

