% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoidalFit.R
\name{sinusoidalFit}
\alias{sinusoidalFit}
\title{Fits a sinusoidal curve to a time-series object}
\usage{
sinusoidalFit(spread.data)
}
\arguments{
\item{spread.data}{A price series for fitting.}
}
\value{
list of values are returned
\item{lastValue}{The last fitted FV}
\item{fitted.values}{The entire range of fitted fair values }
}
\description{
Fits a sinusoidal curve to a time-series object
}
\author{
Helena Ristov and Nicholas Dregne
}

