% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeVolumeVol.R
\name{TradeVolumeVol}
\alias{TradeVolumeVol}
\title{Adjust Volatility based on Liquidity (Daily Traded Volume)}
\usage{
TradeVolumeVol(OR_Contracts, StartTime, EndTime, Types, MinTrdVolume,
  ID = "X.RIC", DataSource = "TRS")
}
\arguments{
\item{OR_Contracts}{List of outright contracts to liquidity test.  Derive other types from OR_Contracts vector.}

\item{StartTime}{Beginning date of liquidity analysis.}

\item{EndTime}{Ending date of liquidity analysis.}

\item{Types}{Types of combinations to examine for liquidity.}

\item{MinTrdVolume}{Minimum amount of liquidity necessary to prevent a liquidity risk adjustment to the regular volatility measurement.}
}
\description{
Calculates Volatility Controllling for Potential Limited Liquidity.
}
\details{
Compares recent daily traded volume to a minimally-accepted daily volume and adjusts volatility if violated.
}
\author{
Nicholas Dregne
}

