% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vol.Imbalance.R
\name{Vol.Imbalance}
\alias{Vol.Imbalance}
\title{Calculates A Volume Imbalance Trigger from Buy and Sell Raw Data}
\usage{
Vol.Imbalance(contract, contract.type, start.date, end.date, incr = list(units
  = "hours", size = 1))
}
\arguments{
\item{contract}{An Instrument identifier}

\item{contract.type}{Types of Syn and Direct}

\item{start.date}{The start time of the evaluation period}

\item{end.date}{The end time of the evaluation period}

\item{incr}{The increment for the volume imbalance rollup. This is a list with units and size parameters}
}
\value{
list of Aroon indicators and triggers:
\item{imb.sell.trigger}{A trigger of 1 when there is abnormal sell activity occurring in the market}
\item{imb.buy.trigger}{A trigger of 1 when there is abnormal buy activity occurring in the market}
}
\description{
Calculates A Volume Imbalance Trigger from Buy and Sell Raw Data
}
\author{
Helena Ristov
}
